+++ saw-central/src/SAWCentral/Crucible/MIR/Builtins.hs
@@ -12,6 +12,8 @@ module SAWCentral.Crucible.MIR.Builtins
     -- ** Setup
     mir_alloc
   , mir_alloc_mut
+  , mir_alloc_raw_ptr_const
+  , mir_alloc_raw_ptr_mut
   , mir_ref_of
   , mir_ref_of_mut
   , mir_assert
@@ -52,6 +54,8 @@ module SAWCentral.Crucible.MIR.Builtins
   , mir_f32
   , mir_f64
   , mir_lifetime
+  , mir_raw_ptr_const
+  , mir_raw_ptr_mut
   , mir_ref
   , mir_ref_mut
   , mir_slice
@@ -164,14 +168,21 @@ ppMIRAbortedResult = ppAbortedResult (\_gp -> mempty)
 -----
 
 mir_alloc :: Mir.Ty -> MIRSetupM SetupValue
-mir_alloc = MIRSetupM . mir_alloc_internal Mir.Immut
+mir_alloc = MIRSetupM . mir_alloc_internal MirPointerRef Mir.Immut
 
 mir_alloc_mut :: Mir.Ty -> MIRSetupM SetupValue
-mir_alloc_mut = MIRSetupM . mir_alloc_internal Mir.Mut
+mir_alloc_mut = MIRSetupM . mir_alloc_internal MirPointerRef Mir.Mut
 
--- | The workhorse for 'mir_alloc' and 'mir_alloc_mut'.
-mir_alloc_internal :: Mir.Mutability -> Mir.Ty -> CrucibleSetup MIR SetupValue
-mir_alloc_internal mut mty = do
+mir_alloc_raw_ptr_const :: Mir.Ty -> MIRSetupM SetupValue
+mir_alloc_raw_ptr_const = MIRSetupM . mir_alloc_internal MirPointerRaw Mir.Immut
+
+mir_alloc_raw_ptr_mut :: Mir.Ty -> MIRSetupM SetupValue
+mir_alloc_raw_ptr_mut = MIRSetupM . mir_alloc_internal MirPointerRaw Mir.Mut
+
+-- | The workhorse for 'mir_alloc', 'mir_alloc_mut', 'mir_alloc_raw_ptr_const',
+-- and 'mir_alloc_raw_ptr_mut'.
+mir_alloc_internal :: MirPointerKind -> Mir.Mutability -> Mir.Ty -> CrucibleSetup MIR SetupValue
+mir_alloc_internal pkind mut mty = do
   st <- get
   let mcc = st ^. Setup.csCrucibleContext
       col = mcc ^. mccRustModule ^. Mir.rmCS ^. Mir.collection
@@ -208,6 +219,7 @@ mir_alloc_internal mut mty = do
   Setup.currentState . MS.csAllocs . at n ?=
     Some (MirAllocSpec { _maConditionMetadata = md
                       , _maType = tpr
+                      , _maPtrKind = pkind
                       , _maMutbl = mut
                       , _maMirType = mty
                       , _maLen = 1
@@ -237,7 +249,7 @@ mir_ref_of_internal label mut val = MIRSetupM $ do
       nameEnv = MS.csTypeNames (st ^. Setup.csMethodSpec)
 
   ty  <- typeOfSetupValue cc env nameEnv val
-  ptr <- mir_alloc_internal mut ty
+  ptr <- mir_alloc_internal MirPointerRef mut ty
 
   tags <- view Setup.croTags
   let md = MS.ConditionMetadata
@@ -633,7 +645,7 @@ mir_postcond term = MIRSetupM $ do
   Setup.crucible_postcond loc term
 
 mir_points_to ::
-  SetupValue {- ^ LHS reference -} ->
+  SetupValue {- ^ LHS reference/pointer -} ->
   SetupValue {- ^ RHS value -} ->
   MIRSetupM ()
 mir_points_to ref val =
@@ -662,14 +674,14 @@ mir_points_to ref val =
               }
      Setup.addPointsTo (MirPointsTo md ref [val])
 
--- | Perform a set of validity checks on the LHS reference value in a
+-- | Perform a set of validity checks on the LHS reference or pointer value in a
 -- 'mir_points_to' command. In particular:
 --
--- * Check that the LHS is in fact a valid reference type.
+-- * Check that the LHS is in fact a valid reference or pointer type.
 --
 -- Returns the 'Mir.Ty' that the LHS points to.
 mir_points_to_check_lhs_validity ::
-  SetupValue {- ^ LHS reference -} ->
+  SetupValue {- ^ LHS reference/pointer -} ->
   W4.ProgramLoc {- ^ The location in the program -} ->
   Setup.CrucibleSetupT MIR TopLevel Mir.Ty
 mir_points_to_check_lhs_validity ref loc =
@@ -680,7 +692,8 @@ mir_points_to_check_lhs_validity ref loc =
      refTy <- typeOfSetupValue cc env nameEnv ref
      case refTy of
        Mir.TyRef referentTy _ -> pure referentTy
-       _ -> throwCrucibleSetup loc $ "lhs not a reference type: "
+       Mir.TyRawPtr referentTy _ -> pure referentTy
+       _ -> throwCrucibleSetup loc $ "lhs not a reference or pointer type: "
                                   ++ show (PP.pretty refTy)
 
 mir_unsafe_assume_spec ::
@@ -899,6 +912,12 @@ mir_f64 = Mir.TyFloat Mir.F64
 
 mir_lifetime :: Mir.Ty
 mir_lifetime = Mir.TyLifetime
+
+mir_raw_ptr_const :: Mir.Ty -> Mir.Ty
+mir_raw_ptr_const ty = Mir.TyRawPtr ty Mir.Immut
+
+mir_raw_ptr_mut :: Mir.Ty -> Mir.Ty
+mir_raw_ptr_mut ty = Mir.TyRawPtr ty Mir.Mut
 
 mir_ref :: Mir.Ty -> Mir.Ty
 mir_ref ty = Mir.TyRef ty Mir.Immut